+++ .gitignore
@@ -3,4 +3,4 @@ sqlcell_dev*.ipynb
 *.pyc
 **/.ipynb_checkpoints
 .DS_Store
-
+*.egg-info
