+++ sqlcell/db.py
@@ -35,3 +35,31 @@ class DBSessionHandler(object):
     
     def dispose(self):
         pass
+
+class EngineHandler(DBSessionHandler):
+    """remove all engines from sqlcell.db:
+       %%sql refresh
+       may have to use @cell_line_magic.
+       add multiple new engines:
+       %%sql add
+       foo=<engine str>
+       bar=<another engine str>
+       baz=<another engine str>"""
+    def __init__(self, *args, **kwargs):
+        super().__init__()
+        
+    def view(self):
+        "show all alias/engines"
+        pass
+    
+    def refresh(self, cell):
+        self.session.query(self.Engines).delete()
+        self.session.commit()
+
+    def alias(self):
+        "allow alias for each engine"
+        pass
+
+    def list(self, *srgs, **kwargs):
+        pass
+
